Merge ColorMapping with Master

+++ Projects/ASI Admin/src/Web Application/Controllers/Show/ExcelUploadController.cs
@@ -1,4 +1,4 @@
-﻿using asi.asicentral.interfaces;
+using asi.asicentral.interfaces;
 using asi.asicentral.model.show;
 using asi.asicentral.services;
 using asi.asicentral.util.show;
@@ -30,19 +30,23 @@ namespace asi.asicentral.web.Controllers.Show
             return View();
         }
 
-        public ShowCompany UpdateShowCompanyData(DataTable ds, int rowId, int showId = 0, bool fasiliateFlag = false)
+        public ShowCompany UpdateShowCompanyData(DataTable ds, int rowId, int showId = 0, bool fasiliateFlag = false, List<ShowEmployeeAttendee> employeeAttendees = null)
         {
             ShowCompany company = null;
-            var asinumber = ds.Rows[rowId]["ASINO"].ToString();
-            var name = ds.Rows[rowId]["Company"].ToString();
-            var memberType = ds.Rows[rowId]["MemberType"].ToString();
+            var asinumber = ds.Rows[rowId]["ASINO"].ToString().Trim();
+            var name = ds.Rows[rowId]["Company"].ToString().Trim();
+            var memberType = ds.Rows[rowId]["MemberType"].ToString().Trim();
             if (fasiliateFlag == true)
             {
-                 company = ObjectService.GetAll<ShowCompany>().FirstOrDefault(item => (item.ASINumber == asinumber && item.Name == name && item.MemberType == memberType));
+                var companies = ObjectService.GetAll<ShowCompany>().Where(item => (item.ASINumber == asinumber)).ToList();
+                var specialCharsPattern = @"[\s,\./\\&\?;=]";
+                var compName = Regex.Replace(name, specialCharsPattern, "");
+                company = companies.FirstOrDefault(item => Regex.Replace(item.Name, specialCharsPattern, "").Equals(compName, StringComparison.CurrentCultureIgnoreCase) && 
+                                                           item.MemberType.Equals(memberType, StringComparison.CurrentCultureIgnoreCase));
             }
             else
             {
-                company = ObjectService.GetAll<ShowCompany>().FirstOrDefault(item => (item.ASINumber == asinumber || (item.Name == name && item.MemberType == memberType)));
+                company = ObjectService.GetAll<ShowCompany>().FirstOrDefault(item => (item.ASINumber == asinumber || (item.Name.Equals(name, StringComparison.CurrentCultureIgnoreCase) && item.MemberType.Equals(memberType, StringComparison.CurrentCultureIgnoreCase))));
             }
             if (company == null)
             {
@@ -139,22 +143,32 @@ namespace asi.asicentral.web.Controllers.Show
             if (company.MemberType == "Distributor" || fasiliateFlag)
             {
                 // update showEmployee
-                var firstName = ds.Rows[rowId]["FirstName"].ToString();
-                var lastName = ds.Rows[rowId]["LastName"].ToString();
+                var firstName = ds.Rows[rowId]["FirstName"].ToString().Trim();
+                var lastName = ds.Rows[rowId]["LastName"].ToString().Trim();
                 string phone = string.Empty;
                 string email = string.Empty;
                 if (ds.Columns.Contains("Phone"))
                 {
-                    phone = ds.Rows[rowId]["Phone"].ToString();
+                    phone = ds.Rows[rowId]["Phone"].ToString().Trim();
                 }
                 if (ds.Columns.Contains("Email Address"))
                 {
-                    email = ds.Rows[rowId]["Email Address"].ToString();
+                    email = ds.Rows[rowId]["Email Address"].ToString().Trim();
                 }
-                if (!string.IsNullOrEmpty(firstName) && !string.IsNullOrEmpty(lastName))
+                if (!string.IsNullOrEmpty(firstName) && !string.IsNullOrEmpty(lastName) )
                 {
                     ShowEmployee employee = null;
-                    employee = company.Employees.FirstOrDefault(item => (item.FirstName == firstName && item.LastName == lastName));
+                    if (!string.IsNullOrEmpty(email))
+                    {
+                        employee = company.Employees.FirstOrDefault(item => !string.IsNullOrEmpty(item.Email) && item.Email.Trim().Equals(email, StringComparison.CurrentCultureIgnoreCase));
+                    }
+                    
+                    if( employee == null)
+                    {
+                        employee = company.Employees.FirstOrDefault(item => (item.FirstName.Trim().Equals(firstName, StringComparison.CurrentCultureIgnoreCase) &&
+                                                                             item.LastName.Trim().Equals(lastName, StringComparison.CurrentCultureIgnoreCase)));
+                    }
+
                     if (employee == null)
                     {
                         employee = new ShowEmployee()
@@ -168,10 +182,8 @@ namespace asi.asicentral.web.Controllers.Show
                     employee.LastName = lastName;
                     employee.EPhone = phone;
                     employee.Email = email;
-                    employee.CreateDate = DateTime.UtcNow;
                     employee.UpdateDate = DateTime.UtcNow;
                     employee.UpdateSource = "ExcelUploadcontroller-Index";
-
 
                     if (fasiliateFlag)
                     {
@@ -214,6 +226,9 @@ namespace asi.asicentral.web.Controllers.Show
 
                         attendee.EmployeeAttendees.Add(employeeAttendee);
                     }
+
+                    if( employeeAttendees != null )
+                        employeeAttendees.Add(employeeAttendee);
                 }
             }
             #endregion update distributor data
@@ -377,6 +392,8 @@ namespace asi.asicentral.web.Controllers.Show
                                     parent.Add((IDictionary<string, object>)pc);
                                 }
                                 DataTable excelDataTable = ToDictionary(parent);
+                                // all employeeAttendees for this event
+                                var employeeAttendees = new List<ShowEmployeeAttendee>();
                                 for (int i = 0; i < excelDataTable.Rows.Count; i++)
                                 {
                                     var memberType = excelDataTable.Rows[i]["MemberType"].ToString();
@@ -407,8 +424,8 @@ namespace asi.asicentral.web.Controllers.Show
                                         return View("../Show/ViewError", objErrors);
                                     }
 
-                                    UpdateShowCompanyData(excelDataTable, i, objShow.Id, fasiliateFlag);
-                                }
+                                    UpdateShowCompanyData(excelDataTable, i, objShow.Id, fasiliateFlag, employeeAttendees);
+                                }                                
 
                                 ObjectService.SaveChanges();
 
@@ -417,30 +434,54 @@ namespace asi.asicentral.web.Controllers.Show
                                 log.Debug("Index - start updating attendee data");
                                 var showAttendees = ObjectService.GetAll<ShowAttendee>().Where(item => item.ShowId == objShow.Id).ToList();
                                 var attendeesToBeDeleted = showAttendees.Where(attendee => attendee.IsExisting == false).ToList();
-                                for (var i = attendeesToBeDeleted.Count() - 1; i >= 0; i--)
+                                if (attendeesToBeDeleted.Count > 0)
                                 {
-                                    var attendee = attendeesToBeDeleted[i];
-                                    if (attendee.EmployeeAttendees != null && attendee.EmployeeAttendees.Any())
+                                    for (var i = attendeesToBeDeleted.Count() - 1; i >= 0; i--)
                                     {
-                                        for (var j = attendee.EmployeeAttendees.Count() - 1; j >= 0; j--)
+                                        var attendee = attendeesToBeDeleted[i];
+                                        if (attendee.EmployeeAttendees != null && attendee.EmployeeAttendees.Any())
                                         {
-                                            ObjectService.Delete(attendee.EmployeeAttendees[j]);
+                                            for (var j = attendee.EmployeeAttendees.Count() - 1; j >= 0; j--)
+                                            {
+                                                ObjectService.Delete(attendee.EmployeeAttendees[j]);
+                                            }
                                         }
+
+                                        if (attendee.DistShowLogos != null && attendee.DistShowLogos.Any())
+                                        {
+                                            for (var j = attendee.DistShowLogos.Count() - 1; j >= 0; j--)
+                                            {
+                                                ObjectService.Delete(attendee.DistShowLogos[j]);
+                                            }
+                                        }
+
+                                        ObjectService.Delete<ShowAttendee>(attendee);
                                     }
 
-                                    if (attendee.DistShowLogos != null && attendee.DistShowLogos.Any())
-                                    {
-                                        for (var j = attendee.DistShowLogos.Count() - 1; j >= 0; j--)
-                                        {
-                                            ObjectService.Delete(attendee.DistShowLogos[j]);
-                                        }
-                                    }
-
-                                    ObjectService.Delete<ShowAttendee>(attendee);
+                                    showAttendees.ForEach(a => a.IsExisting = false);
+                                    ObjectService.SaveChanges();
+                                    log.Debug(string.Format("{0} company attendees have been deleted for '{1}' after uploading", attendeesToBeDeleted.Count, objShow.Name));
                                 }
 
-                                showAttendees.ForEach(a => a.IsExisting = false);
-                                ObjectService.SaveChanges();
+                                // delete any employee attendees not in the sheet
+                                var attendeeIds = ObjectService.GetAll<ShowAttendee>().Where(item => item.ShowId == objShow.Id).Select(a => a.Id).ToList();
+                                var attendees = ObjectService.GetAll<ShowEmployeeAttendee>().Where(e => attendeeIds.Contains(e.AttendeeId)).ToList();
+                                var countDel = 0;
+                                for (var k = attendees.Count - 1; k >= 0; k--)
+                                {
+                                    if (employeeAttendees.FirstOrDefault(a => a.EmployeeId == attendees[k].EmployeeId) == null)
+                                    { // delete employee from attendee list only, not from database
+                                        countDel++;
+                                        ObjectService.Delete(attendees[k]);
+                                    }
+                                }
+
+                                if (countDel > 0)
+                                {
+                                    ObjectService.SaveChanges();
+                                    log.Debug(string.Format("{0} employee attendees have been deleted for '{1}' after uploading", countDel, objShow.Name));
+                                }
+
                                 log.Debug("Index - end updating attendee data - " + (DateTime.Now - postAddingStart).TotalMilliseconds);
 
                             }
